test suites for gitbase + bblfshd(mocked up)
This is a suite that checks if gitbase correctly handles the difficulties that may appear during the interaction with bblfshd GRPC server, such as parse request errors, parse errors, broken UASTs, long parse time evaluation etc.

It uses different bblfshd mockups to be served, regression-core to spin up the gitbase and perform queries.

To assert gitbase behavior it's stderr is being parsed after the query execution.

This tests are been compiled to a simple binary that's added to regression gitbase docker and should be launched during the gitbase regression test routine.

closes #18

Signed-off-by: lwsanty <lwsanty@gmail.com>